refactor: move RelationOneField next to its impls in relation_one.rs
The trait lived in relation.rs while every impl is in relation_one.rs.
Move the definition to sit with its impls; relation.rs keeps only
RelationManyField. The public `schema::RelationOneField` path is unchanged.
